+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## v1.5.1 [unreleased]
 
+### Bugfixes
+
+- [#1938](https://github.com/influxdata/kapacitor/issues/1938): pagerduty2 should use routingKey rather than serviceKey
 - [#1982](https://github.com/influxdata/kapacitor/pull/1982): Fix KafkaTopic not working from TICKscript
 - [#1989](https://github.com/influxdata/kapacitor/pull/1989): Improve Kafka alert throughput.